\name{lsa}
\alias{createLSAspace}
\alias{showLSAspace}
\alias{foldinLSAspace}
\title{Latent Semantic Analysis (LSA)}
\description{
  Creates a latent semantic space from a given document-term matrix,
  displays it, and handles folding-in of additional documents into
  a given latent semantic space.
}
\usage{
createLSAspace( x, dims=dimcalc(method="share") )
showLSAspace( LSAspace )
foldinLSAspace( docvecs, LSAspace )
}
\arguments{
   \item{x}{a textmatrix.}
   \item{dims}{either the number of dimensions or a function.}
   \item{LSAspace}{a latent semantic space generated by createLSAspace.}
   \item{docvecs}{a textmatrix.}
}
\details{
  
LSA combines the classical vector space model -- well known in 
textmining -- with a singular value decomposition (SVD), a two-mode 
factor analysis. Thereby, bag-of-words representations of texts can 
be mapped into a modified vector space that is assumed to reflect 
semantic structure.



}
\value{
  \item{LSAspace}{a latent semantic space (for createLSAspace and foldinLSAspace.}
  \item{textmatrix}{a textmatrix representation of the latent semantic space (for showLSAspace).}
}
\author{ Fridolin Wild \email{fridolin.wild@wu-wien.ac.at} }
\seealso{ \code{\link{textmatrix}} }
\examples{

}
\keyword{algebra}
\keyword{array}
