\name{foldinLSAspace}
\alias{foldinLSAspace}
\title{Ex-post fold-in of documents into a latent semantic space}
\description{
Additional documents can be mapped into the pre-exisiting
latent semantic space without influencing the factor
distribution of the space. Extremely helpful, when
new documents should not influence the existing
latent semantic structure.
}
\usage{
foldinLSAspace( docvecs, LSAspace )
}
\arguments{
   \item{LSAspace}{a latent semantic space generated by createLSAspace.}
   \item{docvecs}{a textmatrix.}
}
\details{

To keep additional documents from influencing the factor distribution
of the underlying text corpus, they can be folded-in after the singular
value decomposition done by \code{createLSAspace()}.

Background Information:
Folding-in is done by calculating a pseudo document vector \code{mi} in 
\code{Mk} of new documents as shown in equation 1 and 2 (cf. Berry et al., 1995):


}
\value{
  \item{textmatrix}{a textmatrix representation of the additional documents in the latent semantic space.}
}
\author{ Fridolin Wild \email{fridolin.wild@wu-wien.ac.at} }
\seealso{ \code{\link{textmatrix}}, \code{\link{createLSAspace}}, \code{\link{showLSAspace}} }
\examples{

# create a first textmatrix with some files
td = tempdir()
dir.create(td)
write( c("dog", "cat", "mouse"), file=paste(td, "data_doc1.dt", sep="/")
write( c("hamster", "mouse", "sushi"), file=paste(td, "data_doc2.dt", sep="/")
write( c("dog", "monster", "monster"), file=paste(td, "data_doc3.dt", sep="/")
matrix1 = textmatrix(td)
unlink(td, recursive=TRUE)

# create a second textmatrix with some different files
td = tempdir()
dir.create(td)
write( c("cat", "mouse", "mouse"), file=paste(td, "data2_doc1.dt", sep="/")
write( c("nothing", "mouse", "monster"), file=paste(td, "data2_doc2.dt", sep="/")
write( c("cat", "monster", "monster"), file=paste(td, "data2_doc3.dt", sep="/")
matrix2 = textmatrix(td)
unlink(td, recursive=TRUE)

# create an LSA space from matrix1
space1 = createLSAspace(matrix1, dims=dimcalc(method="share"))
showLSAspace(space1)

# fold matrix2 into the space generated by matrix1
foldinLSAspace( matrix2, space1)

}
\keyword{algebra}
\keyword{array}
