% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{one_class_mnist}
\alias{one_class_mnist}
\title{One class mnist dataset}
\usage{
one_class_mnist(class, split = "train", valid_split = 0.1, ...)
}
\arguments{
\item{class}{the digit.}

\item{split}{'train', 'valid' or 'test'.}

\item{valid_split}{the fraction of the full training set used for validation.}
}
\description{
Filter a single class of the mnist dataset for training
}
