\name{showLSAspace}
\alias{showLSAspace}
\title{Display a latent semantic space generated by Latent Semantic Analysis (LSA)}
\description{
  Returns a latent semantic space (created by createLSAspace) in
  textmatrix format: rows are terms, columns are documents.
}
\usage{
showLSAspace( LSAspace )
}
\arguments{
   \item{LSAspace}{a latent semantic space generated by createLSAspace.}
}
\details{

To allow comparisons between terms and documents, the internal
format of the latent semantic space needs to be converted to
a classical document-term matrix (just like one generated by
\code{textmatrix()}).

Remark: One of the advantages of LSA is, that terms and documents
can be compared directly. For this, however, you have to access
the partial matrices from an LSAspace directly. Look at (Berry, 1995)
for more information.

}
\value{
  \item{textmatrix}{a textmatrix representation of the latent semantic space.}
}
\author{ Fridolin Wild \email{fridolin.wild@wu-wien.ac.at} }
\seealso{ \code{\link{textmatrix}}, \code{\link{createLSAspace}}, \code{\link{foldinLSAspace}} }
\references{Berry, M., Dumais, S., and O'Brien, G (1995) \emph{Using Linear Algebra for Intelligent Information Retrieval}. In: SIAM Review, Vol. 37(4), pp.573--595. }
\examples{

# create some files
td = tempdir()
dir.create(td)
write( c("dog", "cat", "mouse"), file=paste(td, "data_doc1.dt", sep="/")
write( c("hamster", "mouse", "sushi"), file=paste(td, "data_doc2.dt", sep="/")
write( c("dog", "monster", "monster"), file=paste(td, "data_doc3.dt", sep="/")

# read files into a document-term matrix
myMatrix = textmatrix(td)

# create the latent semantic space
myLSAspace = createLSAspace(myMatrix, dims=dimcalc(method="raw")) 

# display it as a textmatrix again
round(showLSAspace(myLSAspace),2) # should give the original

# create the latent semantic space
myLSAspace = createLSAspace(myMatrix, dims=dimcalc(method="share")) 

# display it as a textmatrix again
myNewMatrix = showLSAspace(myLSAspace) 
myNewMatrix # should look be different!

# compare two terms with the cosine measure
cosine(myNewMatrix["dog",], myNewMatrix["cat",])

# compare two documents with pearson
cor(myNewMatrix[,1], myNewMatrix[,2], method="pearson")

# clean up
unlink(td, recursive=TRUE)

}
\keyword{algebra}
\keyword{array}
