\name{specialchars}
\docType{data}
\alias{specialchars}
\title{List of special character html entities and their character replacement}
\description{
This list contains entities (\code{specialchars$entities}) and their replacement character (\code{specialchars$replacement}), as used by textvector to cleanup html code: for example, this is used to replace the html entity \&auml\; with the character ae. You can use this data set with \code{data(specialchars)}.

}
\usage{
   data(specialchars)
}
\author{ Fridolin Wild \email{f.wild@open.ac.uk}}
\format{list of language specific html entities and their replacement}
\keyword{datasets}
