\name{foldinLSAspace}
\alias{foldinLSAspace}
\title{Ex-post folding-in of textmatrices into an existing latent semantic space}
\description{
Additional documents can be mapped into a pre-exisiting
latent semantic space without influencing the factor
distribution of the space. Applied, when additional documents 
must not influence the calculated existing latent semantic 
factor structure.
}
\usage{
foldinLSAspace( docvecs, LSAspace )
}
\arguments{
   \item{LSAspace}{a latent semantic space generated by createLSAspace.}
   \item{docvecs}{a textmatrix.}
}
\details{

To keep additional documents from influencing the factor distribution
calculated previously from a particular text basis, they can be folded-in 
after the singular value decomposition performed in \code{createLSAspace()}.

Background Information:
For folding-in, a pseudo document vector \code{mi} of the new documents 
is calculated into as shown in the equations (1) and (2) (cf. Berry et al., 1995):

\code{(1) di = t(v) \%*\% Tk \%*\% solve(diag(Sk)) }

\code{(2) mi = Tk \%*\% diag(Sk) \%*\% t(di) }

The document vector \code{t(v)} in equation (1) is identical to an additional 
column of an input textmatrix \code{M} with the term frequencies of the 
essay to be folded-in. \code{Tk} and \code{Sk} are the truncated matrices 
from the SVD applied through \code{createLSAspace()} on a given text 
collection to construct the latent semantic space. The resulting vector
\code{mi} from equation (2) is identical to an additional column in the
textmatrix representation of the latent semantic space (as produced by 
\code{showLSAspace()}).

}
\value{
  \item{textmatrix}{a textmatrix representation of the additional documents in the latent semantic space.}
}
\author{ Fridolin Wild \email{fridolin.wild@wu-wien.ac.at} }
\seealso{ \code{\link{textmatrix}}, \code{\link{createLSAspace}}, \code{\link{showLSAspace}} }
\examples{

# create a first textmatrix with some files
td = tempdir()
dir.create(td)
write( c("dog", "cat", "mouse"), file=paste(td, "D1", sep="/") )
write( c("hamster", "mouse", "sushi"), file=paste(td, "D2", sep="/") )
write( c("dog", "monster", "monster"), file=paste(td, "D3", sep="/") )
matrix1 = textmatrix(td, minWordLength=1)
unlink(td, recursive=TRUE)

# create a second textmatrix with some more files
td = tempdir()
dir.create(td)
write( c("cat", "mouse", "mouse"), file=paste(td, "A1", sep="/") )
write( c("nothing", "mouse", "monster"), file=paste(td, "A2", sep="/") )
write( c("cat", "monster", "monster"), file=paste(td, "A3", sep="/") )
matrix2 = textmatrix(td, vocabulary=rownames(matrix1), minWordLength=1)
unlink(td, recursive=TRUE)

# create an LSA space from matrix1
space1 = createLSAspace(matrix1, dims=dimcalc(method="share"))
showLSAspace(space1)

# fold matrix2 into the space generated by matrix1
foldinLSAspace( matrix2, space1)

}
\keyword{algebra}
\keyword{array}
