\name{triples}
\alias{getTriple}
\alias{setTriple}
\alias{delTriple}
\alias{getSubjectId}
\title{Binding Triples to a Matrix}
\description{
  Allows to store, manage and retrieve SPO-triples (subject, predicate, object)
  bound to the document columns of a document term matrix.
}
\usage{
getTriple ( M, subject, predicate )
setTriple ( M, subject, predicate, object )
delTriple ( M, subject, predicate, object )
getSubjectId ( M, subject )
}
\arguments{
   \item{M}{the document term matrix (see dt\_matrix).}
   \item{subject}{column number or column name (e.g. 'doc3' or 3).}
   \item{predicate}{predicate of the triple sentence (e.g. 'has\_category' or 'has\_grade').}
   \item{object}{value of the triple sentence (e.g. '14' or 14).}
}
\details{
  
  SPO-Triples are simple facts of the uniform structure
  subject + predicate + object. A subject is typically
  a document in the given document term matrix M,
  i.e. its document title (as in the column names) or its 
  column position. A key-value pair (the \option{predicate} 
  and the \option{object}) can be bound to this \option{subject}.
  
  This can be used, for example, to store classification
  information about the documents of the applied text base.
  So whenever you have a new document to add to your
  knowlege base, you can first find similar (=highly
  correlating) documents and offer their category 
  entries then as potential candidates.
  
  The triple data is stored in the environment of M
  which is constructed in the methods \code{dt_matrix()},
  \code{pseudo_docs()} or \code{query()}. 
  
  Whenever you are using a matrix that has not been 
  generated by these functions, you have to add an 
  environment to it by:
  
  \code{environment(mymatrix) = new.env()}
  
  In \code{getTriple()} the arguments \option{subject} and \option{predicate}
  are optional.
  
}
\value{
  \item{dtm}{the document-term-matrix (incl. row- and column-header).}
}
\author{ Fridolin Wild \email{fridolin.wild@wu-wien.ac.at} }
\seealso{ \code{\link{dt_matrix}}, \code{\link{pseudo_docs}}}
\examples{

    x = matrix(2,2,3) # we fake a document term matrix
    rownames(x) = c("dog","mouse") # fake term names
    colnames(x) = c("doc1","doc2","doc3") # fake doc titles
    environment(x) = new.env() # usually done by dt_matrix/pseudo_docs
    
    setTriple(x, "doc1", "has_category", "15")
    setTriple(x, "doc2", "has_category", "7")
    setTriple(x, "doc1", "has_grade", "5")
    setTriple(x, "doc1", "has_category", "11")
    
    getTriple(x, "doc1")
    getTriple(x, "doc1")[[2]]
    getTriple(x, "doc1", "has_category") # -> [1] "15" "11"
    
    delTriple(x, "doc1", "has_category", "15")
    getTriple(x, "doc1", "has_category") # -> [1] "11"
    
}
\keyword{algebra}
\keyword{array}
